Put limits on the simplifier
Summary:
Limit the number of times it can iterate over the entire CFG, and the number of
blocks that it can add to the CFG.

The block limit won't be honored 100%, but it'll be close.  We currently add 3
blocks at a time and checking the limit ahead of generating the blocks is
annoying, so we'll check it afterwards.

Reviewed By: SonicField

Differential Revision: D66045677

fbshipit-source-id: 9918a5695a5784a363505a5ca1479473b4b18c86