DLOG time taken to compile functions and their code size
Summary:
For debug logs, always report the total time taken to compile a function plus
its code size.

Moves PassTimer to Jit/jit_time_log.h and renames it to Timer.  It's not tied to
the concept of a compiler pass, it's just a timestamp that you can ask for
elapsed time.

Reviewed By: SonicField

Differential Revision: D66045854

fbshipit-source-id: 129870b3801be35cc58e2f312dd6311c9796f703