Fix co_localsplus sizing
Summary:
In CPython's code.h it says of `PyCodeObject::co_nlocalsplus`:
```
    int co_nlocalsplus;           /* number of local + cell + free variables */
```
This turns out to be very misleading - it IS the number of those variables but that number is not equal to the sum of the count of each type. The reason for this is some variables can be both a local and a cell.

For more context see: https://github.com/python/cpython/blob/d6b3e78504b3168c432b20002dbcf8ec9a435e61/Objects/codeobject.c#L334

PR up to fix the comment: https://github.com/python/cpython/pull/126993

Reviewed By: martindemello

Differential Revision: D66140761

fbshipit-source-id: 635e6d01616ac27c56af8159590283ab5ffbf0e1