Fix unlock bug in ThreadedCompileContext + refactors
Summary:
Fix a bug where unlock() unconditionally set the owning thread to the default
value of std::thread::id.  This is wrong because the mutex is recursive, the
unlocking thread might still be the owner.  The value of the thread ID was used
by canAccessSharedData(), where it could have produced a false negative and
triggered invalid assertion failures.  In order to know whether to unset the
owning thread this needs to track how many levels of recursion it has collected.

Otherwise this tries to add comments + use type aliases a little more.  I
suspect we'll need to move this from BorrowedRef<> to Ref<> in the future once
we think about free-threading.

THREADED_COMPILE_SERIALIZED_CALL is unused.

Reviewed By: jbower-fb

Differential Revision: D66209281

fbshipit-source-id: dd719a69a64a71271fd2a287b63b170581257531