Check if functions are JIT compiled via the CodeAllocator
Summary:
Instead of having to refer to the global jit::Context object to check if a
PyFunctionObject is compiled, the caller can ask the CodeAllocator if the
function's vectorcall is contained inside of it.  Makes dependencies much
simpler.

This new isJitCompiled() function needs to be callable from C for Static Python.
The best place I found to put it is in Jit/compiled_function.h, but I had to add
some #ifdefs.

Reviewed By: jbower-fb

Differential Revision: D63651176

fbshipit-source-id: a8081b2f6acf16aea14770814e85d790700f3a5a