Make sure UpdatePrevInstr pass is enabled and points to correct code data
Summary:
We weren't adding the `UpdatePrevInstr` instructions most of the time as the pass was only enabled for some native tests. I've also modified the pass to handle not having valid line-table data. Not sure this will ever come up in practice, but we did have one runtime test with this issue

Also, the code data we need to point to is `co_code_adaptive` (not the cached deopted copy),  as this is what the interpreter should actually resume at for continued execution on deopt.

Reviewed By: alexmalyshev

Differential Revision: D66240745

fbshipit-source-id: 530417589ce578ced6fd30e7cc439306edf38c43