Stop specializing vectorcalls to Python objects
Summary:
The JIT is burning in addresses to the vectorcall entry points of Python
callables.  This sounds good in that it lets us skip the extra overhead of going
through `PyObject_Vectorcall`.  However it also means that calls to callables
that haven't been JIT-compiled yet will always get routed through the
interpreter.

In theory we should always be registering all existing functions with the JIT, because
of the use of `PyUnstable_GC_VisitObjects()` in the CinderX module initialization.
However, that won't pick up any functions that have been migrated to the permanent
generation via `gc.freeze()`. So if we manage to compile calls to those functions
before we compile them themselves then they'll never get to call their JIT-compiled
code.

Loading the vectorcall entry point dynamically means we'll pick up any changes
made by the JIT.  In the future, we can explore tracking these calls and
patching them with JIT-compiled functions.

Reviewed By: jbower-fb

Differential Revision: D63651175

fbshipit-source-id: 8e1929591cbd21150a94a0917b915c3769a8cf34