Include RETURN_CONST in list of consts that are trimmed from the end
Summary: There's a list of const opcodes which are derived by the converters which are known to use the const converter. Because `RETURN_CONST` is just an optimization applied after emitting there are no registered opcodes for it. This just adds the additional opcode into the list for 3.12 so that we don't trim constants used by the 3.10 trim consts pass.

Reviewed By: alexmalyshev

Differential Revision: D66380878

fbshipit-source-id: bec96f7a6f81b0c5958130b42afae56ff38cc708