Fix LOAD_GLOBAL and add PUSH_NULL optimization
Summary:
In 3.12 LOAD_GLOBAL has the offset of the global to be loaded offset by 1, with the low bit indicating that `LOAD_GLOBAL` should also push a NULL value on the stack. There's an optimization that turns `PUSH_NULL` + `LOAD_GLOBAL` into a single `LOAD_GLOBAL` with that opcode.

This implements that optimization and updates our conversion for `LOAD_GLOBAL`.

Reviewed By: martindemello

Differential Revision: D66369756

fbshipit-source-id: c79dcd3443ebb79f8137538c822cb552dff7931e