Add 3.12 handlers and don't create dict on every opcode
Summary:
D66369756 needs to add a new handler for `PUSH_NULL`. So this sets up `FlowGraphOptimizer312` to have its own set of handlers.

Currently we're re-creating the handler dictionary on every single opcode which is going to be unnecessarily inefficient. This moves to creating the dictionary once and explicitly passing `self` to the handler.

Reviewed By: martindemello

Differential Revision: D66508953

fbshipit-source-id: 4ebb97ba0e110b337d9f920627309ecda23b109d