Remove functions from jit::Context that forward to jit::CompiledFunction
Summary:
Most of these functions boil down to:

  T foo(BorrowedRef<PyFunctionObject> func) {
    CompiledFunction* compiled_func = lookupFunc(func);
    return compiled_func != nullptr ? compiled_func->foo() : T{};
  }

and they only have a single callsite (always in pyjit.cpp) so we might as well
inline them.  The one that's different is inlinedFunctionsStats() and that one
definitely shouldn't be in jit::Context given the amount of work it spends
constructing Python objects.

Reviewed By: jbower-fb

Differential Revision: D66662299

fbshipit-source-id: d64b49d3efc9c6922854e898fa8d00924199e897