Fix simple await code generation
Summary:
Our code generation for a simple method that does an await such as:
```
async def foo():
    await a
```
Is a little bit off.  First, we mis-compute the stack depth of a generator function, this is because we have the stack effect for `RETURN_GENERATOR` incorrect.

Next we're inserting `SETUP_CLEANUP` in at the wrong spot - this should go near the beginning of the method, not in the current basic block, this is an issue because we mis-calculate the stack depth for the `YIELD_VALUE`.

Finally, our stack depth in `label_exception_targets` ends up being incorrect, because we don't see the `SEND` block as a stand-alone basic block which has fall-through, therefore our calculation of the exception stack ends up being incorrect.

Reviewed By: martindemello

Differential Revision: D66548051

fbshipit-source-id: 9a26c4dfd6f32a2c5a674ad9f6774691085c642e