Support disabling and re-enabling the JIT
Summary:
Add a new `deopt_all` parameter to `cinderjit.disable()`.  This will force all
currently compiled functions back to the interpreter.  It defaults to `False`.

Add a new `cinderjit.enable()` function that will scan through deopted functions
and re-attach them to existing compiled code objects.

As part of these changes, the `force_compile()` function will now respect
whether the JIT is enabled or not, and will become a nop when the JIT is
disabled.

Reviewed By: SonicField

Differential Revision: D66662323

fbshipit-source-id: 21481d4351813bbd18b899dceac9c421f9dd0551