Fix HIR parser and strobelight tests for 3.12
Summary:
The HIR parser needs to learn about LoadEvalBreaker and RunPeriodicTasks.  The
strobelight tests are checking for defined offset symbols.  Given that the rest
of strobelight support doesn't work yet for 3.12, I say we turn the test into a
nop with UPGRADE_NOTE instead of UPGRADE_ASSERT.  We won't lose this as long as
we keep cinderx/Upgrade/upgrade_assert.h around.

Reviewed By: SonicField

Differential Revision: D66730122

fbshipit-source-id: 2a94a067a010c9697c7f0e448f8154c851e0e3e5