Fix indexes in generated code for cells/locals
Summary:
Fixes up some issues surrounding our locals indexes. First, in our disassembly we need to populate the code objects `localsplusnames` so that we can print the names successfully when disassembling.

Another issue is that we're generally not updating `ioparg` in the passes that fix things up, or when we're initially populating the instructions. When updating the instructions we only update `oparg` if it was already an int.

Finally when we're doing `self.varnames.index(varname)` we actually create the varname it it doesn't exist, messing up the generated code object's varnames. We should check if it's present first.

Finally `LOAD_FROM_DICT_OR_DEREF` needs to be using `_convert_DEREF` instead of `_convert_NAME` to get the right index.

Reviewed By: martindemello

Differential Revision: D66666459

fbshipit-source-id: 538d479dcf7b3f5cb7542ae45f60a49757819ff0