Generalize deopt for generators a little
Summary: This is primarily for 3.12+ where we'll be short-circuiting deopt for generators in some places. I.e. deopting directly on operations like `throw()` and `close()` whereas today we implicitly deopt by throwing an exception into the generator.  Avoiding the need to ever throw into a JIT generator (only to have it immediately deopt) keeps things simpler in 3.12.

Reviewed By: alexmalyshev

Differential Revision: D66567766

fbshipit-source-id: c8a987cf904b997bf50617bd063d2ecf93a3e5c0