Pull out a cinder CodeGenerator base class/mixin.
Summary:
Start the process of porting the cinderx and static python compilers to 3.12.

The basic idea is that the code generation will differ based on python version,
but the AST nodes that we want to specialise for cinder should be the same in
both cases, so we define a mixin providing those overrides and delegating to
methods in concrete subclasses.

Note that CinderCodeGenBase is not a "pure" mixin; it inherits from the base
CodeGenerator (largely for type checking purposes), and does bring in all the
hassles of multiple inheritance. Identifying the exact behaviour cinder
overrides from 3.10 should make it easier to modify the design later, if we
decide we want to go a different route.

Reviewed By: alexmalyshev

Differential Revision: D66841326

fbshipit-source-id: 84826947e312152fa4766b910324f85adbfde7c2