Consistently use set_to_nop
Summary:
Currently we're not consistently using set_to_nop, which means we're not clearing ioparg, and we end up with some EXTENDED_ARG nops. This just uses set_to_nop everywhere.

In 3.10 sometimes CPython doesn't clear `ioparg` either, which leads to `EXTENDED_ARG` `NOP` sequences, so we need to handle this on a version by version basis.

Reviewed By: martindemello

Differential Revision: D66898525

fbshipit-source-id: cd7b1c76d94f4e6b0af635b6a1322986b7fd28f8