Fix issue with shadowing global in inlined list comp
Summary:
There seems to be a lingering issue w/ list comprehensions when an inlined list comp has a variable that is being shadowed from the global scope turn into a cell var. In that case we end up treating it like a cell in the global scope too.

We need to not add these variables to cells on the module, and instead, let the generation of the cell-related opcodes to fill in the code object.

Reviewed By: alexmalyshev

Differential Revision: D67111736

fbshipit-source-id: fac23b7d0d7875ae19ff847021a82b75587ffa34