Match CPython's handling of __classdict__ and __class__
Summary: We've got some hacky support to match CPython's `__class__` and `__classdict__`, but with inlined comprehensions this ends up behaving differently, so we need to more closely match it. The problem is that when we include these vars in `cellvars` we end up emitting different loads/stores inside of a class comprehension.

Reviewed By: alexmalyshev

Differential Revision: D67108368

fbshipit-source-id: 7db1c0b0cfa1e9bda9cadb39538d94710236465b