Make sbs test sharding strategy actually not still include a long run of all tests
Summary:
We have a strategy of sharding the sbs tests so that they can run in parallel, and we don't end up with a run that takes a really long time.

Unfortunately this strategy is being defeated by us leaving the original tests in the test_sbs_stdlib test which runs serially.

This updates us to not publish the classes as attributes on the module and therefore we only have the tests in the shared test libraries.

This is a little bit less convenient for doing runs on dev servers, but that's infrequent and this line can always be brought back for that purpose.

Reviewed By: jbower-fb

Differential Revision: D67205978

fbshipit-source-id: 4112834c19002051fd6df4f90990b2c7354a7ad5