Support tracking call counts in 3.12
Summary:
Was waiting on the bundled 3.12 interpreter loop to land, now we can initialize
code object's extra data and bump up the count on each function call.

Needed to make changes to `cinderx/Common/code.h` to be able to use it across C
and C++ code.  It also tweaks `initCodeExtra()` to return the `CodeExtra` object
directly, so the interpreter doesn't have to issue a following `codeExtra()` to
read it.

This enables JitAuto mode in 3.12.

Reviewed By: DinoV

Differential Revision: D67068783

fbshipit-source-id: 7f82ecfaae63b4b709c811ccfacd683c93f3c490