Implement INVOKE_FUNCTION opcode
Summary:
Implements the INVOKE_FUNCTION opcode minus the adaptive support.

The one interesting thing here is the usage of the `args[invoke_function_args(frame->f_code->co_consts, oparg)]` for the incoming arguments.

This works fine for now for the basic opcode, but we may want to move to encoding the number of the arguments in the oparg.

We actually kind of support this as we always generate these as EXTENDED_ARG 0/INVOKE_FUNCTION so that we have space to encode the args when we turn it into an INVOKE_FUNCTION_CACHED. We just don't emit the number of args w/ the initial opcode.

Reviewed By: martindemello

Differential Revision: D67174973

fbshipit-source-id: 1d8e9ef8d3396de9c8fc8a819f422f6552e0a36c