Add a cinderx.jit module
Summary:
It's tedious to always do the song and dance of:

```lang=python
try:
    from cinderjit import is_jit_compiled
except ImportError
    def is_jit_compiled(func):
        return False
```

because `cinderjit` only shows up when the JIT is enabled.  So build it into the
cinderx module itself.  That way anybody who includes CinderX as an extension
should be able to run JIT-related methods, even if it is disabled.

This also allows us to implement JIT-related functionality in pure Python, see the
following diff.

In the future I'd like to see this replace `cinderjit` entirely. It's an odd behavior
with how we sometimes automagically import `cinderjit` after importing `cinderx`.

Reviewed By: czardoz

Differential Revision: D67220109

fbshipit-source-id: e87a968eb319f5ff58084edc5db9f3dcaf83f5fe