Add a context manager for pausing the JIT
Summary:
Closer to what users would expect from Python for this kind of functionality.
Manually using disable() and enable() is tedious because you need to consider
exceptions.

Note that this only exists in `cinderx.jit`, not in `cinderjit`. `cinderjit` is a Native
extension and it's far easier to implement this in pure Python.

Reviewed By: czardoz

Differential Revision: D67220110

fbshipit-source-id: 0893ed2aba3b5e64d1edd47e7d285e43a39ec58a