Allow EXTENDED_ARG with 0 argument
Summary:
We generate an EXTENDED_ARG with a zero argument for INVOKE_FUNCTION and INVOKE_METHOD.

We may want to revisit what we do there, but to keep things simple between 3.10 and 3.12 we should keep doing this. This just overrides the instruction so that we allow the 0 oaprg version. It also removes an assertion which raises when they're already present in the instruction stream.

Reviewed By: martindemello

Differential Revision: D67310870

fbshipit-source-id: 61e7d8e624be12bc0cfd89282a5574a84bc5dc57