Do not assert that non-JIT generators are not executing on shutdown
Summary:
Apparently we can be in a situation where interpeter generators have frames which are marked as executing on Python shutdown. This sounds broken to me but maybe the interpreter just doesn't care to mark the frames as non-executing on shutdown?

In any case, we only need to deopt JIT generators so are only concerned if those are still running.

Reviewed By: alexmalyshev

Differential Revision: D67415509

fbshipit-source-id: 60627cab2d3d75d18b2f4a5978a463996cbe1695