Fixup CallCFunc calls for 3.12
Summary:
Fix up the uses of the HIR `CallCFunc` instruction for calls related to coroutines: `_PyGen_yf` and `_PyCoro_GetAwaitableIter`. In 3.12 these can't call the CPython internal implementations anymore even with borrowing as we need to modify them to deal with the new generator types we've introduced.

This diff puts the horse a little before the cart in that it has a commented out references to `JitCoro_CheckExact()` which will be introduced in a following change. I've left these in as I think it makes it clearer why we need a new implementation here.

Reviewed By: DinoV

Differential Revision: D67425131

fbshipit-source-id: 53751a02a66efc01f4ea3a093b8d9fa81a63c493