Fix issues around class construction in 3.12 static
Summary:
Currently we can define a class, but it doesn't really end up being static, because we're using the normal BUILD_CLASS.

We need to instead use the invoke to the static build class. This then fails because we don't have PyHeapType_GET_MEMBERS. But that's now just a call to PyObject_GetItemData.

We don't yet support type params on static classes, but I've filed a task to add that - T210915794. It's not hard, we just need to include .type_params as part of the bases.

We also need to fix up a usage of Py_IDENTIFIER.

Reviewed By: alexmalyshev

Differential Revision: D67405886

fbshipit-source-id: 542965f8a1a35fb60ebd0cdbd0e64c8a363406d6