Give compiler signature information for typed methods unavailable in 3.12
Summary:
In 3.10 we can reflect against the methods of chklist and chkdict and provide errors on calling them.

In 3.12 we don't have the typed method metadata, so we need to embed it in the compiler.

This information is just generated from inspecting the 3.10 metadata which is just the dictionaries seen here.

Reviewed By: alexmalyshev

Differential Revision: D67461122

fbshipit-source-id: b3f63ed2e817b2e0f5c2e04a248a29aee3665343