Provide better error message when value has been deleted
Summary:
We have a test case that works in 3.10 due to shadow code (test_cinderx.test_compiler.test_static.patch.StaticPatchTests.test_patch_final_bad_ret_del).

In 3.12 we try the lookup after the value has been deleted, and provide an error which says it has no value. But we can fallback to the originals and provide a message that it has been removed.

Reviewed By: jbower-fb

Differential Revision: D67505959

fbshipit-source-id: 51610431994f2c0fd7524bf2cc58e0d79d3e83c9