add immortalize_heap
Summary:
We borrow a few chunks from the upstream GC module, and implement an `immortalize_heap()` function.

For 3.10, we call into the gc module for the actual implementation, and avoid having to borrow any stuff. I only added that to ensure compatibility, nothing uses the cinderx immortalize APIs at all.

Reviewed By: alexmalyshev

Differential Revision: D66996583

fbshipit-source-id: a91b56c340835a7e5ee71a8a03872f953b7203b2