Provide better error message when value has been deleted
Summary:
Reverting the revert (D67555337)

We have a test case that works in 3.10 due to shadow code (test_cinderx.test_compiler.test_static.patch.StaticPatchTests.test_patch_final_bad_ret_del).

In 3.12 we try the lookup after the value has been deleted, and provide an error which says it has no value. But we can fallback to the originals and provide a message that it has been removed.

Reviewed By: martindemello

Differential Revision: D67603437

fbshipit-source-id: e7024785d8a0424579a77dcc87973ccd24053c55