Handle duplicate preloaders after preloading
Summary:
Reading through the code here, we first check for an existing preloader, if it's
not there then we make a new one and add it onto the global preloader manager.
There's an assertion in the preloader manager that we never try to add
duplicates.

Making a preloader has the potential to run Python code, which means we can
re-enter the JIT compiler and thus the preloader.  So after we've returned from
creating our new preloader, we need to check again that there's no existing
preloader.

Reviewed By: brittanyrey

Differential Revision: D67601122

fbshipit-source-id: 1361f8cb532599db6f957c24715d9a7f37b9db22