Make v-table entrypoints more well formed
Summary:
In S477539 there was a crash, but the stack trace was getting lost at the v-table thunk.

Currently our v-table entry points are doing the normal prologue after the jump to the real static entry point. Even though we maintain the frame pointer chain after the jump lldb seems rather persnickety about this and it loses the stack trace.

We should instead follow a normal function prologue and then lldb will successfully be able to follow the stack trace. This just involves moving the push/and rbp update at the beginning of the function, removing some nops to account for the new bytes, and then adjusting the jump offset.

Reviewed By: alexmalyshev

Differential Revision: D67654947

fbshipit-source-id: 5b9836be54d68de901c646114f16fbb83ffd2c8c