Use Callable instead of FunctionType in JIT API
Summary:
I tried to use FunctionType because that only applies to Python functions and
not to builtin functions implemented natively, e.g.

```lang=python
>>> from types import FunctionType
>>> def foo(): pass
...
>>> isinstance(foo, FunctionType)
True
>>> isinstance(id, FunctionType)
False
```

However Pyre doesn't support FunctionType, see

[Pyre Playground](https://pyre-check.org/play?input=%23%20pyre-strict%0A%0Afrom%20types%20import%20FunctionType%0Afrom%20typing%20import%20Any%2C%20Callable%2C%20TypeVar%0A%0ATRet%20%3D%20TypeVar(%22TRet%22)%0A%0Adef%20takes_func(f%3A%20FunctionType)%20-%3E%20None%3A%0A%20%20%20%20pass%0A%0A%0Adef%20takes_callable(f%3A%20Callable%5B...%2C%20TRet%5D)%20-%3E%20None%3A%0A%20%20%20%20pass%0A%0A%0Adef%20nop()%20-%3E%20None%3A%0A%20%20%20%20pass%0A%0A%0Adef%20inc(n%3A%20int)%20-%3E%20int%3A%0A%20%20%20%20return%20n%20%2B%201%0A%0A%0Atakes_func(nop)%0Atakes_callable(nop)%0A%0Atakes_func(inc)%0Atakes_callable(inc))

and the `types` module just isn't as helpful as the `typing` module is for
users.  So switch to using `Callable` instead.  This means that our static
typing will accept more kinds of functions, including those we don't want.  The
actual implementation will raise an exception when it receives something that
isn't a Python function, so the runtime behavior would be exactly the same.

Reviewed By: DinoV

Differential Revision: D67872264

fbshipit-source-id: c5698dab45dd75ebdd60bee6e0eb5d8b553893a3