Make v-table entrypoints more well formed
Summary:
Reverting the revert: D67876917 - there were some IG failures that coincidently showed up with this diff, but weren't due to this diff.

Currently our v-table entry points are pushing rbp/rsp after a jmp. Even though we maintain the frame pointer chain after the jump this ends up confusing lldb and we lose the stack trace.

We should instead follow a normal function prologue and then we'll successfully be able to follow the stack trace.

Reviewed By: alexmalyshev

Differential Revision: D67880212

fbshipit-source-id: 8e38b81caf8a7a15061f304a13e34bb1b37ddd39