Add cinderx.jit.lazy_compile()
Summary:
Primarily meant for testing the behavior of the JIT.

This function is different than force_compile() in that it registers the
function to be compiled the first time it gets called.  This is closer to how
the JIT operates when used with a JIT list, and it also mimicks -X jit-auto=0.

Reviewed By: martindemello

Differential Revision: D68223949

fbshipit-source-id: 6cd1ee44840e28c3a5f58c0b33017503fdeccd16