Fix precompile_all() not running with multiple workers
Summary:
When calling `precompile_all()` the expectation is that the number of workers
will be determined by the `PYTHONJITBATCHCOMPILEWORKERS` env var.  If that's not
set, then the default is to run the compilation entirely within the current
thread.

Turns out `precompile_all()` was accidentally setting the default number of
workers to 1, which looked like the user had overwritten the number by calling
`precompile_all(workers=1)`.  The result was that the env var was silently being
ignored. It's supposed to default to `workers=0`, which means "get the worker count
from the env var else 1".

Reviewed By: martindemello

Differential Revision: D68230337

fbshipit-source-id: df66dfb9e7f4355834533bdd9db41fb590f1ae1b