Handle a single source of truth for argument hydration
Summary:
Currently we store the original function and each v-table entry figures out how to get the original code object back in a different way. This unifies so we always store the code object so they can all grab it in the same way.

Ultimately this will let us simplify our stubs because we won't need to have a whole bunch of entry points which will vary based upon how we get back to the function object to figure out what the native arguments are.

This adds a single signature object which captures the information for the signature of an object. For most signatures we'll be able to just use a baked in shared signature because typically we have no primitive values and we don't have too many arguments.

For signatures with primitives or more arguments then we'll allocate a signature object and use that.

Reviewed By: alexmalyshev

Differential Revision: D67728210

fbshipit-source-id: 59e91caf037111dd4415b471570bbe556606db91