Support importing `cinderx` without `_cinderx`
Summary:
Trying to make importing the top-level module not an error if the runtime isn't
compatible with the underlying `_cinderx` module. This way we can use the
`cinderx` module in libraries that are shared across `use_cinder=True` and
`use_cinder=False` binaries.

Added a `cinderx.is_initialized()` function so users can check if CinderX was
loaded and initialized properly.  Capture the import error too to make errors
easier to debug.

Reviewed By: czardoz, SonicField

Differential Revision: D68346117

fbshipit-source-id: c7a64a513e66488de9227862e310063d23d30557