Allow using cinderx as a real cpp_python_extension
Summary:
Essentially, the implementation of the plan outlined in https://fb.workplace.com/groups/pyupgrades/posts/3520818281555632

I reworked our internal build rules so that 3.10 and 3.12 have entirely separate
build graphs. We use the right version of each dependency by using `select()` with
the right python version constraint.

This allows us to clean up most of the complications caused by the `pyverchoose`
macro, which required us to always specify the "desired" version in the form of a
buckconfig flag.

Now, we have a standard "cpp_python_extension" for CinderX,
and load it up as a regular Python dependency, on binaries that have `use_cinder` set
to `True` (and have `py_version` set to `3.12`).

In this diff, I haven't touched how 3.10 works - it's essentially unchanged. I've also
_not_ refactored all the test scripts to get rid of `cinder.use_3_12=true`, instead,
our custom `cpp_binary` and `cpp_unittests` still require that flag. We can get rid of
it in a separate change, where we directly build `xyz_3.10` or `xyz_3.12`

Reviewed By: alexmalyshev

Differential Revision: D67169281

fbshipit-source-id: 13747e8d88d03f21cfabc214ad5a0d96e905ab6e