Don't import types or typing from cinderx
Summary:
Otherwise this will break Lib/test/test_site.py, see the new comment in
cinderx/__init__.py for more context.

I could have worked around this in test_site.py for 3.10.cinder, but
eventually we'd need to do the same thing for 3.12 and I'd rather keep
that as close to upstream as possible.

Reviewed By: yilei

Differential Revision: D68559940

fbshipit-source-id: a33399ac9b20056cf1840ed8b84eb7b27cf5122c