Unify on using code in arg hydration
Summary:
Reland of D67728210.

The issue here was that `_PyClassLoader_GetThunkSignatureFromCode` was updating the argument info using `i` (`sig->ta_argtype[i + extra_args] = arg_type;`) but the actual argument that is primitive needs to be determined using the `type_descr` as it specifies the index. Updated it to do:
```
      int arg_pos = _PyClassLoader_GetArgumentDescrPosition(checks, i);
      sig->ta_argtype[arg_pos + extra_args] = arg_type;
```

And added a new test case to invoke.py.

Currently we store the original function and each v-table entry figures out how to get the original code object back in a different way. This unifies so we always store the code object so they can all grab it in the same way.

Ultimately this will let us simplify our stubs because we won't need to have a whole bunch of entry points which will vary based upon how we get back to the function object to figure out what the native arguments are.

This adds a single signature object which captures the information for the signature of an object. For most signatures we'll be able to just use a baked in shared signature because typically we have no primitive values and we don't have too many arguments.

For signatures with primitives or more arguments then we'll allocate a signature object and use that.

Reviewed By: pilleye

Differential Revision: D68335517

fbshipit-source-id: 542df37144d1583cb0a0742c043322a6d6701a5b