Hackily fix shutdown issues when cinderx functions are compiled
Summary:
Turns out that JIT-compiling `cinderx.strictify_static()` will prevent the
cinderx module from being finalized on shutdown.  This is a silent issue on
3.10.  On 3.12 it turns into an ASAN error, as the interned strings table is
deallocated in `Py_FinalizeEx()` but then accessed by the global destructor in
the `GlobalCacheManager` singleton.

Smells like a reference cycle to me, but I'm not sure where it's coming from
yet, so let's patch over it for now.

Reviewed By: martindemello

Differential Revision: D68557970

fbshipit-source-id: dbc32a23fda707049bbe114967ef787dc70b3906