Completely remove pyverchoose
Summary:
Use the right constraint modifiers in our top-level targets, so we never have to rely on pyverchoose.

This has a bunch of changes for our team, internally. The devx significantly changes, and going forward, all top-level targets where we used to use the `use_3_12` will have to instead use the right `_3.12` target as needed.

Reviewed By: alexmalyshev

Differential Revision: D68747392

fbshipit-source-id: 186db3ffce40c9d9bc0d4c53b4838220e2670cef