Only use `vte_state` for vectorcall entry
Summary: We no longer need to use the entry point defined in the v-table to invoke functions when we're going through vectorcall because all of our state fields are valid vectorcallable objects. This let's us directly vectorcall them from the interpreter and also let's use drop the `InvokeMethod` HIR instruction which can simply be replaced by a `VectorCall`.

Reviewed By: alexmalyshev

Differential Revision: D67951787

fbshipit-source-id: 9d0d4b670a0694e7563a54822e7e0a60a5b22c9f