Add class method thunk
Summary:
Removes the tuple based state for class methods and uses a specific class method thunk.

This is starting to show how we're getting better separation of concerns. Our handling of co-routines and class-methods is now orthogonal, with the thunk handling the fixup's to `self` and the outer thunk handling the co-routine type checking behavior.

Reviewed By: alexmalyshev

Differential Revision: D67728207

fbshipit-source-id: 3809f1d5e5464890962ceeb2032d6158168fced2