Add static method thunk
Summary:
Similar to other changes in this stack this adds a thunk for calling static functions replacing the existing tuple state. This is vectorcall-able unlike the tuple.

While this is the "static method" thunk it can actually be used for anything that needs to drop `self`, and so we'll use it for those purposes elsewhere as when we have an instance value that we're calling we won't drop self.

Reviewed By: alexmalyshev

Differential Revision: D67728206

fbshipit-source-id: 58931602f56ac8f1a369a08994de47e20f327391