Don't raise an error when using types like typing.NamedTuple as annotations
Summary:
There are a handful of types like `NamedTuple` and `TypedDict` that are meant
to create specialised types when subclassed. It is technically incorrect to use
them directly as annotations, however raising the error is undesirable from a
user point of view and we therefore simply convert them to dynamic types.

Reviewed By: DinoV

Differential Revision: D68806571

fbshipit-source-id: 76c278fc613a8038beba04770ab9ff8e0dc80f35