Fix some refcounting issues in the 3.12 interpreter loop
Summary:
We have some issues w/ ref counting in the interpreter loop, in particular around when we encounter an error.

It turns out `ERROR_IF` is subtly different than `goto error;`.

`ERROR_IF` will handle the stack reference adjustments in the generated interpreter loop, but it does those without decref'ing any of the inputs.

`goto error;` will have no stack reference changes, and so the references will still be on the stack, which it will decref and cleanup.

So we need to tweak things so that we get the right things decref'd.

Reviewed By: martindemello

Differential Revision: D68745648

fbshipit-source-id: 44a0adb2d8daa168a864160fd6442926dc0fe796