Don't use tp_cache for StaticGeneric
Summary:
Currently we use `tp_cache` for the StaticGeneric class to ensure we get unique values. Unfortuantely this causes us to leak StaticGeneric, which actually has leaking the entire module, and that causes us to leak `_static` as well. This makes it impossible to get the module clear event at the end of `_static`'s lifetime which we need to not hit an assert in the runtime that static built-in classes are de-initialized in the proper order in 3.12: https://www.internalfb.com/code/fbsource/[4897531466186ec24045dcdaa30aa2fa4741c631]/third-party/python/3.12/Objects/typeobject.c?lines=5027

We don't currently really use `StaticGeneric` anywhere except for tests, so I was tempted to delete it entirely, but we can get the right behavior by using weakref's ourselves to track the lifetime and remove the type from the cache if any of the type vars go away.

Reviewed By: itamaro

Differential Revision: D69078503

fbshipit-source-id: 90f4f883a14813e9344477b64ff4dc85bf5a47af